requirements.txt
python-telegram-bot>=22.7,<23.0
pymongo>=4.13,<5.0
httpx>=0.28,<1.0
